corrreção na migration de createPoints

+++ back-end/src/database/migrations/20200604195655_createPoints.ts
@@ -7,8 +7,8 @@ export async function up(knex: Knex): Promise<void> {
     table.string('name').notNullable();
     table.string('email').notNullable();
     table.string('whatsapp').notNullable();
-    table.decimal('latitude').notNullable();
-    table.decimal('longitude').notNullable();
+    table.float('latitude').notNullable();
+    table.float('longitude').notNullable();
     table.string('city').notNullable();
     table.string('uf', 2).notNullable();
   });
